For some unknown reason, when we changed the interface of the piece selector algorithm to return a Vec<PieceSelection> instead of a HashMap<u32, Peer>, performance decreased absurdly: it went from about 100MiB/s download speed to at most 1.25 MiB/s. After lots of debugging, we found out that the problem was caused by the fact we were now requesting the pieces in order (piece 1, then piece 2, 3, and so on) and not out of order as before (due to the fact that the HashMap does not guarantee order).
This commit shuffles the selection before returning it, guaranteeing that it won't sorted. After that, performance went back to the expected levels. The reason why requesting pieces in order decreases performance is above our comprehension, but one hyphotesis is that we may be generating some sort of concurrence or race condition in our peer (aria2c), causing it to take longer to answer us. Another hyphotesis is that we may take longer to write sequential blocks to the disk, which would also decrease the overall performance (but for me, that doesn't make sense, since sequential writes should be faster).

Anyways, that's a good thing to mention in our final report and to discuss with Olivier Alphand.